+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMAdministrativeUser Resource
 - Added Compare-MultipleCompares to the ResourceHelper
 - Added CMDistributionGroup Resource
+- Added CMSiteSystemServer Resource
 
 ### Changed
 
@@ -61,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added UDP 1434 to the defaults for the xSCCMPreReqs.
 - Fixed newline in the CMIniFile resource.
 - Removed WSUS top level feature.
+- Added Security Scopes to CMDistributionGroup Resource
 - Added SiteSystems, SiteSystemsToInclude, and SiteSystemsToExclude and SecurityScopes,
   SecurityScopesToInclude, SecurityScopesToExclude to CMBoundaryGroup resource.